\name{SPEM-package}
\alias{SPEM-package}
\docType{package}
\title{
S-system parameter estimation method package
}
\description{
The function in this package allows for the computation of parameters in the n-gene S-system from time series data.
}
\details{
\tabular{ll}{
Package: \tab SPEM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-01-13\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}


}
\author{
Yang, X-Y, Dent, J. E. and Nardini, C.

Maintainer: Xinyi Yang \email{yangxinyi@picb.ac.cn}
}
\keyword{Package}


